moves game to components from pages
