+++ components/game.js
